Move bout notification content to its own file
